largest-contentful-paint/observe-css-generated-text.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <script>
    function checkText(entry, expectedSize, expectedID, beforeRender) {
      assert_equals(entry.entryType, 'largest-contentful-paint', 'Entry should be of type largest-contentful-paint');
      assert_greater_than_equal(entry.renderTime, beforeRender, 'Render time should be greater than time just before rendering');
      assert_greater_than_equal(performance.now(), entry.renderTime, 'renderTime should be less than current time');
      assert_equals(entry.startTime, entry.renderTime, 'startTime should equal renderTime');
      assert_equals(entry.duration, 0, 'duration should be 0');
      assert_greater_than_equal(entry.size, expectedSize, 'Size should match expected size');
      assert_equals(entry.loadTime, 0, 'loadTime should be zero');
      assert_equals(entry.id, expectedID, 'ID should match expected ID');
      assert_equals(entry.url, '', 'URL should be empty');
      assert_equals(entry.element, document.getElementById(expectedID), 'Entry element is expected element');
    }


    promise_test(t => {
      return new Promise(resolve => {
        assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
        const observer = new PerformanceObserver(resolve);
        observer.observe({ type: 'largest-contentful-paint' });
        beforeRender = performance.now();
        const element = document.createElement('p');
        element.id = 'css-generated-text';
        document.body.appendChild(element);
      }).then(entryList => {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        checkText(entry, 1200, 'css-generated-text', beforeRender);
      });
    }, "CSS generated text is observable as a LargestContentfulPaint candidate");

    promise_test(t => {
      return new Promise(resolve => {
        assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
        const observer = new PerformanceObserver(resolve);
        observer.observe({ type: 'largest-contentful-paint' });
        beforeRender = performance.now();
        const element = document.createElement('p');
        element.id = 'css-generated-text-attr';
        element.setAttribute('data-text', 'This is some text generated using content:attr() via css');
        document.body.appendChild(element);
      }).then(entryList => {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        checkText(entry, 1200, 'css-generated-text-attr', beforeRender);
      });
    }, "Text generated with CSS using content:attr() is observable as a LargestContentfulPaint candidate");

    promise_test(t => {
      return new Promise(resolve => {
        assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
        const observer = new PerformanceObserver(resolve);
        observer.observe({ type: 'largest-contentful-paint' });
        beforeRender = performance.now();
        const element = document.createElement('span');
        element.id = 'css-generated-text-inline-elem';
        document.body.appendChild(element);
      }).then(entryList => {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        checkText(entry, 1200, 'css-generated-text-inline-elem', beforeRender);
      });
    }, "CSS generated text on a inline element is observable as a LargestContentfulPaint candidate");
  </script>
  <style>
    #css-generated-text::before {
      content: 'This is some text generated via css';
      font-size: 12px;
    }

    #css-generated-text-attr::before {
      content: attr(data-text);
      font-size: 13px;
    }

    #css-generated-text-inline-elem::before {
      content: 'This is some more text generated via css that should be displayed via a span tag';
      font-size: 14px;
    }
  </style>
</body>
